requirements.txt
flask
flask-cors
pandas
torch
openai
google-genai
Pillow
numpy
lxml
beautifulsoup4
